requirements.txt
blinker==1.6.2
click==8.1.4
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Flask==2.3.2
flask-sqlalchemy==3.0.5
fonttools==4.40.0
greenlet==2.0.2
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.24.4
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
PyMySQL==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.18
threadpoolctl==3.1.0
typing-extensions==4.7.1
tzdata==2023.3
Werkzeug==2.3.6
wordcloud==1.9.2
zipp==3.16.0
